pyproject.toml
[tool.poetry]
name = "wallet-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "wallet_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.0"
python-dotenv = "^0.21.0"
sqlalchemy = "^1.4.46"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
